requirements.txt
aniso8601==8.0.0
asgiref==3.2.10
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
djangorestframework==3.12.1
django-heroku==0.3.1
Flask==1.1.2
Flask-RESTful==0.3.8
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
imageio==2.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==2.2.5
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
opencv-python==3.4.9.33
pascal-voc-writer==0.1.4
Pillow==5.4.1
protobuf==3.11.3
psycopg2==2.7.5
psycopg2-binary==2.8.5
PTable==0.9.2
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-json-logger==0.1.8
PyWavelets==1.1.1
requests==2.23.0
requests-toolbelt==0.9.1
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
simplejson==3.17.0
six==1.14.0
-e git+https://github.com/supervisely/supervisely.git@2188283e075481ec3ade6f61f4f5f5164aadc271#egg=supervisely
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.11.2
pytz==2020.1
sqlparse==0.4.1
